+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @coherent.js/runtime
 
+## 1.0.0
+
+### Patch Changes
+
+- @coherent.js/client@1.0.0
+- @coherent.js/core@1.0.0
+- @coherent.js/web-components@1.0.0
+
 ## 1.0.0-beta.3
 
 ### Patch Changes